test: overhauling the test suite to make it easier to maintain
