+++ __tests__/index.test.js
@@ -1,22 +1,15 @@
 const extensions = require('@readme/oas-extensions');
 const Oas = require('oas').default;
-const path = require('path');
-const datauri = require('datauri');
 const oasToHar = require('../src');
 const toBeAValidHAR = require('jest-expect-har').default;
 
 const petstore = require('@readme/oas-examples/3.0/json/petstore.json');
-const commonParameters = require('./__fixtures__/common-parameters.json');
-const multipartFormData = require('./__fixtures__/multipart-form-data.json');
-const multipartFormDataArrayOfFiles = require('./__fixtures__/multipart-form-data/array-of-files.json');
 const serverVariables = require('./__fixtures__/server-variables.json');
 
 expect.extend({ toBeAValidHAR });
 
-const oas = new Oas({});
-
-test('should output a har object', async () => {
-  const har = oasToHar(oas);
+test('should output a HAR object if no operation information is supplied', async () => {
+  const har = oasToHar(new Oas({}));
 
   await expect(har).toBeAValidHAR();
   expect(har).toStrictEqual({
@@ -65,21 +58,25 @@ test('should accept an Operation instance as the operation schema', async () => 
   });
 });
 
-test('should uppercase the method', async () => {
-  const har = oasToHar(oas, { path: '/', method: 'get' });
-
-  expect(har.log.entries[0].request.method).toBe('GET');
-  await expect(har).toBeAValidHAR();
-});
-
 describe('url', () => {
   it('should be constructed from oas.url()', () => {
-    expect(oasToHar(oas, { path: '', method: 'get' }).log.entries[0].request.url).toBe(oas.url());
+    const spec = new Oas(petstore);
+    const operation = spec.operation('/pet', 'post');
+    const har = oasToHar(spec, operation);
+
+    expect(har.log.entries[0].request.url).toBe(`${spec.url()}/pet`);
   });
 
-  // TODO this should probably happen within the Operation class
   it('should replace whitespace with %20', () => {
-    expect(oasToHar(oas, { path: '/path with spaces', method: '' }).log.entries[0].request.url).toBe(
+    const spec = new Oas({
+      paths: {
+        '/path with spaces': {
+          get: {},
+        },
+      },
+    });
+
+    expect(oasToHar(spec, spec.operation('/path with spaces', 'get')).log.entries[0].request.url).toBe(
       'https://example.com/path%20with%20spaces'
     );
   });
@@ -143,1522 +140,49 @@ describe('url', () => {
 
   describe('proxy url', () => {
     const proxyOas = new Oas({
+      paths: {
+        '/path': {
+          get: {},
+        },
+      },
       [extensions.PROXY_ENABLED]: true,
     });
 
     it('should not be prefixed with without option', () => {
-      expect(oasToHar(proxyOas, { path: '/path', method: 'get' }).log.entries[0].request.url).toBe(
-        'https://example.com/path'
-      );
+      const har = oasToHar(proxyOas, proxyOas.operation('/path', 'get'));
+      expect(har.log.entries[0].request.url).toBe('https://example.com/path');
     });
 
     it('should be prefixed with try.readme.io with option', () => {
-      expect(
-        oasToHar(proxyOas, { path: '/path', method: 'get' }, {}, {}, { proxyUrl: true }).log.entries[0].request.url
-      ).toBe('https://try.readme.io/https://example.com/path');
+      const har = oasToHar(proxyOas, proxyOas.operation('/path', 'get'), {}, {}, { proxyUrl: true });
+      expect(har.log.entries[0].request.url).toBe('https://try.readme.io/https://example.com/path');
     });
-  });
-});
-
-describe('parameters', () => {
-  describe('path', () => {
-    it('should pass through unknown path params', () => {
-      expect(oasToHar(oas, { path: '/param-path/{id}', method: '' }).log.entries[0].request.url).toBe(
-        'https://example.com/param-path/id'
-      );
-
-      expect(
-        oasToHar(oas, {
-          path: '/param-path/{id}',
-          method: 'get',
-          parameters: [
-            {
-              name: 'something-else',
-              in: 'path',
-              required: true,
-            },
-          ],
-        }).log.entries[0].request.url
-      ).toBe('https://example.com/param-path/id');
-    });
-
-    it.each([
-      [
-        'should not error if empty object passed in for values',
-        {
-          parameters: [{ name: 'id', in: 'path', required: true }],
-        },
-        {},
-        'https://example.com/param-path/id',
-      ],
-      [
-        'should use default if no value',
-        {
-          parameters: [{ name: 'id', in: 'path', required: true, schema: { default: '123' } }],
-        },
-        {},
-        'https://example.com/param-path/123',
-      ],
-      [
-        'should add path values to the url',
-        {
-          parameters: [{ name: 'id', in: 'path', required: true }],
-        },
-        { path: { id: '456' } },
-        'https://example.com/param-path/456',
-      ],
-      [
-        'should add falsy values to the url',
-        {
-          parameters: [{ name: 'id', in: 'path', required: true }],
-        },
-        { path: { id: 0 } },
-        'https://example.com/param-path/0',
-      ],
-    ])('%s', async (_, operation, formData, expectedUrl) => {
-      const har = oasToHar(
-        oas,
-        {
-          path: '/param-path/{id}',
-          method: 'get',
-          ...operation,
-        },
-        formData
-      );
-
-      await expect(har).toBeAValidHAR();
-
-      expect(har.log.entries[0].request.url).toStrictEqual(expectedUrl);
-    });
-  });
-
-  describe('query', () => {
-    it.each([
-      [
-        'should not add on empty unrequired values',
-        {
-          parameters: [{ name: 'a', in: 'query' }],
-        },
-      ],
-      [
-        'should not add the parameter name as a value if required but missing',
-        {
-          parameters: [{ name: 'a', in: 'query', required: true }],
-        },
-      ],
-      [
-        'should set defaults if no value provided but is required',
-        {
-          parameters: [{ name: 'a', in: 'query', required: true, schema: { default: 'value' } }],
-        },
-        {},
-        [{ name: 'a', value: 'value' }],
-      ],
-      [
-        'should pass in value if one is set and prioritise provided values',
-        {
-          parameters: [{ name: 'a', in: 'query', required: true, schema: { default: 'value' } }],
-        },
-        { query: { a: 'test' } },
-        [{ name: 'a', value: 'test' }],
-      ],
-      [
-        'should add falsy values to the querystring',
-        {
-          parameters: [{ name: 'id', in: 'query' }],
-        },
-        { query: { id: 0 } },
-        [{ name: 'id', value: '0' }],
-      ],
-      [
-        'should handle null array values',
-        {
-          parameters: [{ name: 'id', in: 'query' }],
-        },
-        { query: { id: [null, null] } },
-        [{ name: 'id', value: '&id=' }],
-      ],
-      [
-        'should handle null values',
-        {
-          parameters: [{ name: 'id', in: 'query' }],
-        },
-        { query: { id: null } },
-        [{ name: 'id', value: 'null' }],
-      ],
-      [
-        'should handle null default values',
-        {
-          parameters: [
-            {
-              name: 'id',
-              in: 'query',
-              required: true,
-              schema: {
-                type: 'array',
-                items: {
-                  type: 'string',
-                },
-                default: [null, null],
-              },
-            },
-          ],
-        },
-        { query: {} },
-        [{ name: 'id', value: '&id=' }],
-      ],
-    ])('%s', async (_, operation = {}, formData = {}, expectedQueryString = []) => {
-      const har = oasToHar(
-        oas,
-        {
-          path: '/query',
-          method: 'get',
-          ...operation,
-        },
-        formData
-      );
-
-      await expect(har).toBeAValidHAR();
-      expect(har.log.entries[0].request.queryString).toStrictEqual(expectedQueryString);
-    });
-
-    describe('URI encoding', () => {
-      const spec = new Oas({
-        servers: [{ url: 'https://httpbin.org/' }],
-        paths: {
-          '/anything': {
-            get: {
-              parameters: [
-                { name: 'stringPound', in: 'query', schema: { type: 'string' } },
-                { name: 'stringPound2', in: 'query', schema: { type: 'string' } },
-                { name: 'stringHash', in: 'query', schema: { type: 'string' } },
-                { name: 'stringArray', in: 'query', schema: { type: 'string' } },
-                { name: 'stringWeird', in: 'query', schema: { type: 'string' } },
-                { name: 'array', in: 'query', schema: { type: 'array', items: { type: 'string' } } },
-              ],
-            },
-          },
-        },
-      });
-
-      it('should encode query parameters', async () => {
-        const formData = {
-          query: {
-            stringPound: 'something&nothing=true',
-            stringHash: 'hash#data',
-            stringArray: 'where[4]=10',
-            stringWeird: 'properties["$email"] == "testing"',
-            array: [
-              encodeURIComponent('something&nothing=true'), // This is already encoded so it shouldn't be double encoded.
-              'nothing&something=false',
-              'another item',
-            ],
-          },
-        };
-
-        const operation = spec.operation('/anything', 'get');
-
-        const har = oasToHar(spec, operation, formData);
-        await expect(har).toBeAValidHAR();
-
-        expect(har.log.entries[0].request.queryString).toStrictEqual([
-          { name: 'stringPound', value: 'something%26nothing%3Dtrue' },
-          { name: 'stringHash', value: 'hash%23data' },
-          { name: 'stringArray', value: 'where%5B4%5D%3D10' },
-          {
-            name: 'stringWeird',
-            value: 'properties%5B%22%24email%22%5D%20%3D%3D%20%22testing%22',
-          },
-          { name: 'array', value: 'something%26nothing%3Dtrue&array=nothing%26something%3Dfalse&array=another%20item' },
-        ]);
-      });
-
-      it('should not double encode query parameters that are already encoded', async () => {
-        const formData = {
-          query: {
-            stringPound: encodeURIComponent('something&nothing=true'),
-            stringHash: encodeURIComponent('hash#data'),
-            stringArray: encodeURIComponent('where[4]=10'),
-            stringWeird: encodeURIComponent('properties["$email"] == "testing"'),
-            array: [
-              'something&nothing=true', // Should still encode this one eventhrough the others are already encoded.
-              encodeURIComponent('nothing&something=false'),
-              encodeURIComponent('another item'),
-            ],
-          },
-        };
-
-        const operation = spec.operation('/anything', 'get');
-
-        const har = oasToHar(spec, operation, formData);
-        await expect(har).toBeAValidHAR();
-
-        expect(har.log.entries[0].request.queryString).toStrictEqual([
-          { name: 'stringPound', value: 'something%26nothing%3Dtrue' },
-          { name: 'stringHash', value: 'hash%23data' },
-          { name: 'stringArray', value: 'where%5B4%5D%3D10' },
-          {
-            name: 'stringWeird',
-            value: 'properties%5B%22%24email%22%5D%20%3D%3D%20%22testing%22',
-          },
-          { name: 'array', value: 'something%26nothing%3Dtrue&array=nothing%26something%3Dfalse&array=another%20item' },
-        ]);
-      });
-    });
-  });
-
-  describe('cookie', () => {
-    it.each([
-      [
-        'should not add on empty unrequired values',
-        {
-          parameters: [{ name: 'a', in: 'cookie' }],
-        },
-      ],
-      [
-        'should not add the parameter name as a value if required but missing',
-        {
-          parameters: [{ name: 'a', in: 'cookie', required: true }],
-        },
-      ],
-      [
-        'should set defaults if no value provided but is required',
-        {
-          parameters: [{ name: 'a', in: 'cookie', required: true, schema: { default: 'value' } }],
-        },
-        {},
-        [{ name: 'a', value: 'value' }],
-      ],
-      [
-        'should pass in value if one is set and prioritize provided values',
-        {
-          parameters: [{ name: 'a', in: 'cookie', required: true, schema: { default: 'value' } }],
-        },
-        { cookie: { a: 'test' } },
-        [{ name: 'a', value: 'test' }],
-      ],
-      [
-        'should add falsy values to the cookies',
-        {
-          parameters: [{ name: 'id', in: 'cookie' }],
-        },
-        { cookie: { id: 0 } },
-        [{ name: 'id', value: '0' }],
-      ],
-    ])('%s', async (_, operation = {}, formData = {}, expectedCookies = []) => {
-      const har = oasToHar(
-        oas,
-        {
-          path: '/',
-          method: 'get',
-          ...operation,
-        },
-        formData
-      );
-
-      await expect(har).toBeAValidHAR();
-
-      expect(har.log.entries[0].request.cookies).toStrictEqual(expectedCookies);
-    });
-  });
-
-  describe('header', () => {
-    it.each([
-      [
-        'should not add on empty unrequired values',
-        {
-          parameters: [{ name: 'a', in: 'header' }],
-        },
-      ],
-      [
-        'should not add the parameter name as a value if required but missing',
-        {
-          parameters: [{ name: 'a', in: 'header', required: true }],
-        },
-      ],
-      [
-        'should set defaults if no value provided but is required',
-        {
-          parameters: [{ name: 'a', in: 'header', required: true, schema: { default: 'value' } }],
-        },
-        {},
-        [{ name: 'a', value: 'value' }],
-      ],
-      [
-        'should pass in value if one is set and prioritise provided values',
-        {
-          parameters: [{ name: 'a', in: 'header', required: true, schema: { default: 'value' } }],
-        },
-        { header: { a: 'test' } },
-        [{ name: 'a', value: 'test' }],
-      ],
-      [
-        'should pass accept header if endpoint expects a content back from response',
-        {
-          parameters: [{ name: 'a', in: 'header', required: true, schema: { default: 'value' } }],
-          responses: {
-            200: {
-              content: {
-                'application/xml': { type: 'array' },
-                'application/json': { type: 'array' },
-              },
-            },
-          },
-        },
-        {},
-        [
-          { name: 'Accept', value: 'application/xml' },
-          { name: 'a', value: 'value' },
-        ],
-      ],
-      [
-        'should only add one accept header',
-        {
-          responses: {
-            200: {
-              content: {
-                'application/xml': {},
-              },
-            },
-            400: {
-              content: {
-                'application/json': {},
-              },
-            },
-          },
-        },
-        {},
-        [{ name: 'Accept', value: 'application/xml' }],
-      ],
-      [
-        'should only receive one accept header if specified in values',
-        {
-          parameters: [{ name: 'Accept', in: 'header' }],
-          responses: {
-            200: {
-              content: {
-                'application/json': {},
-                'application/xml': {},
-              },
-            },
-          },
-        },
-        { header: { Accept: 'application/xml' } },
-        [{ name: 'Accept', value: 'application/xml' }],
-      ],
-      [
-        'should add accept header if specified in formdata',
-        {
-          responses: {
-            200: {
-              content: {
-                'application/json': {},
-                'application/xml': {},
-              },
-            },
-          },
-        },
-        { header: { Accept: 'application/xml' } },
-        [{ name: 'Accept', value: 'application/xml' }],
-      ],
-      [
-        'should add falsy values to the headers',
-        {
-          parameters: [{ name: 'id', in: 'header' }],
-        },
-        { header: { id: 0 } },
-        [{ name: 'id', value: '0' }],
-      ],
-    ])('%s', async (_, operation = {}, formData = {}, expectedHeaders = []) => {
-      const har = oasToHar(
-        oas,
-        {
-          path: '/header',
-          method: 'get',
-          ...operation,
-        },
-        formData
-      );
-
-      await expect(har).toBeAValidHAR();
-
-      expect(har.log.entries[0].request.headers).toStrictEqual(expectedHeaders);
-    });
-  });
-});
-
-describe('requestBody', () => {
-  describe('body values', () => {
-    it('should not add on empty unrequired values', () => {
-      const pathOperation = {
-        path: '/body',
-        method: 'get',
-        requestBody: {
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  a: {
-                    type: 'string',
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      expect(oasToHar(oas, pathOperation).log.entries[0].request.postData).toBeUndefined();
-    });
-
-    // TODO extensions[SEND_DEFAULTS]
-    it.skip('should set defaults if no value provided but is required', () => {
-      expect(
-        oasToHar(oas, {
-          path: '/body',
-          method: 'get',
-          requestBody: {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  required: ['a'],
-                  properties: {
-                    a: {
-                      type: 'string',
-                    },
-                  },
-                },
-              },
-            },
-          },
-        }).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify({ a: 'value' }));
-    });
-
-    it('should pass in value if one is set and prioritise provided values', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    required: ['a'],
-                    properties: {
-                      a: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { a: 'test' } }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify({ a: 'test' }));
-    });
-
-    it('should work for RAW_BODY primitives', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      RAW_BODY: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { RAW_BODY: 'test' } }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify('test'));
-    });
-
-    it('should work for RAW_BODY json', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      RAW_BODY: {
-                        type: 'string',
-                        format: 'json',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { RAW_BODY: '{ "a": 1 }' } }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify({ a: 1 }));
-    });
-
-    it('should return empty for falsy RAW_BODY primitives', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      RAW_BODY: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { RAW_BODY: '' } }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify(''));
-    });
-
-    it('should work for RAW_BODY objects', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      RAW_BODY: {
-                        type: 'object',
-                        properties: {
-                          a: {
-                            type: 'string',
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { RAW_BODY: { a: 'test' } } }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify({ a: 'test' }));
-    });
-
-    it('should return empty for RAW_BODY objects', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      RAW_BODY: {
-                        type: 'object',
-                        properties: {
-                          a: {
-                            type: 'string',
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { RAW_BODY: {} } }
-        ).log.entries[0].request.postData.text
-      ).toBeUndefined();
-    });
-
-    it('should return nothing for undefined body property', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      a: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { a: undefined } }
-        ).log.entries[0].request.postData.text
-      ).toBeUndefined();
-    });
-
-    it('should work for schemas that require a lookup', () => {
-      expect(
-        oasToHar(
-          new Oas({
-            components: {
-              requestBodies: {
-                schema: {
-                  content: {
-                    'application/json': {
-                      schema: { type: 'object', properties: { a: { type: 'integer' } } },
-                    },
-                  },
-                },
-              },
-            },
-          }),
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              $ref: '#/components/requestBodies/schema',
-            },
-          },
-          { body: { a: 123 } }
-        ).log.entries[0].request.postData.text
-      ).toStrictEqual(JSON.stringify({ a: 123 }));
-    });
-
-    it('should work for schemas that require a parameters lookup', () => {
-      expect(
-        oasToHar(
-          new Oas({
-            components: {
-              parameters: {
-                authorization: {
-                  name: 'Authorization',
-                  in: 'header',
-                },
-              },
-            },
-          }),
-          {
-            method: 'get',
-            parameters: [
-              {
-                $ref: '#/components/parameters/authorization',
-              },
-            ],
-          },
-          { header: { Authorization: 'test' } }
-        ).log.entries[0].request.headers[0].value
-      ).toBe('test');
-    });
-
-    it('should work for top level primitives', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'post',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'string',
-                  },
-                },
-              },
-            },
-          },
-          { body: 'string' }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify('string'));
-
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'post',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'integer',
-                    format: 'int64',
-                  },
-                },
-              },
-            },
-          },
-          { body: 123 }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify(123));
-
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'post',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'boolean',
-                  },
-                },
-              },
-            },
-          },
-          { body: true }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify(true));
-    });
-
-    it('should work for top level falsy primitives', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'post',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'string',
-                  },
-                },
-              },
-            },
-          },
-          { body: '' }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify(''));
-
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'post',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'integer',
-                    format: 'int64',
-                  },
-                },
-              },
-            },
-          },
-          { body: 0 }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify(0));
-
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'post',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'boolean',
-                  },
-                },
-              },
-            },
-          },
-          { body: false }
-        ).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify(false));
-    });
-
-    it('should not include objects with undefined sub properties', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      a: {
-                        type: 'object',
-                        properties: {
-                          b: {
-                            type: 'string',
-                          },
-                          c: {
-                            type: 'object',
-                            properties: {
-                              d: {
-                                type: 'string',
-                              },
-                            },
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: { a: { b: undefined, c: { d: undefined } } } }
-        ).log.entries[0].request.postData.text
-      ).toBeUndefined();
-    });
-
-    // When we first render the form, formData.body is undefined
-    // until something is typed into the form. When using anyOf/oneOf
-    // if we change the schema before typing anything into the form,
-    // then onChange is fired with `undefined` which causes
-    // this to error
-    it('should not error if `formData.body` is undefined', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      a: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { body: undefined }
-        ).log.entries[0].request.postData
-      ).toBeUndefined();
-    });
-
-    describe('content types', () => {
-      describe('multipart/form-data', () => {
-        let owlbert;
-
-        beforeAll(async () => {
-          owlbert = await datauri(path.join(__dirname, '__fixtures__', 'owlbert.png'));
-
-          // Doing this manually for now until when/if https://github.com/data-uri/datauri/pull/29 is accepted.
-          owlbert = owlbert.replace(';base64', `;name=${encodeURIComponent('owlbert.png')};base64`);
-        });
-
-        it('should handle multipart/form-data request bodies', () => {
-          const fixture = new Oas(multipartFormData);
-          const har = oasToHar(fixture, fixture.operation('/anything', 'post'), {
-            body: { orderId: 12345, userId: 67890, documentFile: owlbert },
-          });
-
-          expect(har.log.entries[0].request.headers).toStrictEqual([
-            { name: 'Content-Type', value: 'multipart/form-data' },
-          ]);
-
-          expect(har.log.entries[0].request.postData).toStrictEqual({
-            mimeType: 'multipart/form-data',
-            params: [
-              { name: 'orderId', value: '12345' },
-              { name: 'userId', value: '67890' },
-              {
-                contentType: 'image/png',
-                fileName: 'owlbert.png',
-                name: 'documentFile',
-                value: owlbert,
-              },
-            ],
-          });
-        });
-
-        it('should handle multipart/form-data request bodies where the filename contains parentheses', () => {
-          // Doing this manually for now until when/if https://github.com/data-uri/datauri/pull/29 is accepted.
-          const specialcharacters = owlbert.replace(
-            'name=owlbert.png;',
-            `name=${encodeURIComponent('owlbert (1).png')};`
-          );
-
-          const fixture = new Oas(multipartFormData);
-          const har = oasToHar(fixture, fixture.operation('/anything', 'post'), {
-            body: { orderId: 12345, userId: 67890, documentFile: specialcharacters },
-          });
-
-          expect(har.log.entries[0].request.headers).toStrictEqual([
-            { name: 'Content-Type', value: 'multipart/form-data' },
-          ]);
-
-          expect(har.log.entries[0].request.postData).toStrictEqual({
-            mimeType: 'multipart/form-data',
-            params: [
-              { name: 'orderId', value: '12345' },
-              { name: 'userId', value: '67890' },
-              {
-                contentType: 'image/png',
-                fileName: encodeURIComponent('owlbert (1).png'),
-                name: 'documentFile',
-                value: specialcharacters,
-              },
-            ],
-          });
-        });
-
-        it('should handle a multipart/form-data request where files are in an array', () => {
-          const fixture = new Oas(multipartFormDataArrayOfFiles);
-          const har = oasToHar(fixture, fixture.operation('/anything', 'post'), {
-            body: {
-              documentFiles: [owlbert, owlbert],
-            },
-          });
-
-          expect(har.log.entries[0].request.postData).toStrictEqual({
-            mimeType: 'multipart/form-data',
-            params: [
-              {
-                name: 'documentFiles',
-                value: JSON.stringify([owlbert, owlbert]),
-              },
-            ],
-          });
-        });
-      });
-
-      describe('image/png', () => {
-        it('should handle a image/png request body', async () => {
-          let owlbert = await datauri(path.join(__dirname, '__fixtures__', 'owlbert.png'));
-
-          // Doing this manually for now until when/if https://github.com/data-uri/datauri/pull/29 is accepted.
-          owlbert = owlbert.replace(';base64', `;name=${encodeURIComponent('owlbert.png')};base64`);
-
-          const har = oasToHar(
-            oas,
-            {
-              path: '/',
-              method: 'post',
-              responses: {
-                200: {
-                  description: 'OK',
-                },
-              },
-              requestBody: {
-                content: {
-                  'image/png': {
-                    schema: {
-                      type: 'string',
-                      format: 'binary',
-                    },
-                  },
-                },
-              },
-            },
-            { body: owlbert }
-          );
-
-          await expect(har).toBeAValidHAR();
-
-          // The `postData` contents here should be the data URL of the image for a couple reasons:
-          //
-          //  1. The HAR spec doesn't have support for covering a case where you're making a PUT request to an endpoint
-          //    with the contents of a file, eg. `curl -T filename.png`. Since there's no parameter name, as this is
-          //    the entire content of the payload body, we can't promote this up to `postData.params`.
-          //  2. Since the HAR spec doesn't have support for this, neither does the `httpsnippet` module, which we
-          //    couple with this library to generate code snippets. Since that doesn't have support for `curl -T
-          //    filename.png` cases, the only thing we can do is just set the data URL of the file as the content of
-          //    `postData.text`.
-          //
-          //  It's less than ideal, and code snippets for these kinds of operations are going to be extremely ugly, but
-          //  there isn't anything we can do about it.
-          expect(har.log.entries[0].request.postData.mimeType).toBe('image/png');
-          expect(har.log.entries[0].request.postData.text).toBe(`${owlbert}`);
-        });
-      });
-    });
-
-    describe('format: `json`', () => {
-      it('should work for refs that require a lookup', () => {
-        expect(
-          oasToHar(
-            new Oas({
-              components: {
-                requestBodies: {
-                  schema: {
-                    content: {
-                      'application/json': {
-                        schema: {
-                          string: 'object',
-                          properties: { a: { type: 'string', format: 'json' } },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            }),
-            {
-              path: '/body',
-              method: 'get',
-              requestBody: {
-                $ref: '#/components/requestBodies/schema',
-              },
-            },
-            { body: { a: '{ "b": 1 }' } }
-          ).log.entries[0].request.postData.text
-        ).toBe(JSON.stringify({ a: JSON.parse('{ "b": 1 }') }));
-      });
-
-      it('should leave invalid JSON as strings', () => {
-        expect(
-          oasToHar(
-            oas,
-            {
-              path: '/body',
-              method: 'post',
-              requestBody: {
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      required: ['a'],
-                      properties: {
-                        a: {
-                          type: 'string',
-                          format: 'json',
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-            { body: { a: '{ "b": invalid json' } }
-          ).log.entries[0].request.postData.text
-        ).toBe(JSON.stringify({ a: '{ "b": invalid json' }));
-      });
-
-      it('should parse valid arbitrary JSON request bodies', () => {
-        expect(
-          oasToHar(
-            oas,
-            {
-              path: '/body',
-              method: 'post',
-              requestBody: {
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'string',
-                      format: 'json',
-                    },
-                  },
-                },
-              },
-            },
-            { body: '{ "a": { "b": "valid json" } }' }
-          ).log.entries[0].request.postData.text
-        ).toBe('{"a":{"b":"valid json"}}');
-      });
-
-      it('should parse invalid arbitrary JSON request bodies as strings', () => {
-        expect(
-          oasToHar(
-            oas,
-            {
-              path: '/body',
-              method: 'post',
-              requestBody: {
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'string',
-                      format: 'json',
-                    },
-                  },
-                },
-              },
-            },
-            { body: '{ "a": { "b": "valid json } }' }
-          ).log.entries[0].request.postData.text
-        ).toBe(JSON.stringify('{ "a": { "b": "valid json } }'));
-      });
-
-      it('should parse valid JSON as an object', () => {
-        expect(
-          oasToHar(
-            oas,
-            {
-              path: '/body',
-              method: 'post',
-              requestBody: {
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      required: ['a'],
-                      properties: {
-                        a: {
-                          type: 'string',
-                          format: 'json',
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-            { body: { a: '{ "b": "valid json" }' } }
-          ).log.entries[0].request.postData.text
-        ).toBe(JSON.stringify({ a: JSON.parse('{ "b": "valid json" }') }));
-      });
-
-      it('should parse one valid JSON format even if another is invalid', () => {
-        expect(
-          oasToHar(
-            oas,
-            {
-              path: '/body',
-              method: 'post',
-              requestBody: {
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      required: ['a'],
-                      properties: {
-                        a: {
-                          type: 'string',
-                          format: 'json',
-                        },
-                        b: {
-                          type: 'string',
-                          format: 'json',
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-            { body: { a: '{ "z": "valid json" }', b: 'invalid json' } }
-          ).log.entries[0].request.postData.text
-        ).toBe(JSON.stringify({ a: { z: 'valid json' }, b: 'invalid json' }));
-      });
-
-      it('should parse one valid JSON format even if another is left empty', () => {
-        expect(
-          oasToHar(
-            oas,
-            {
-              path: '/body',
-              method: 'post',
-              requestBody: {
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      required: ['a'],
-                      properties: {
-                        a: {
-                          type: 'string',
-                          format: 'json',
-                        },
-                        b: {
-                          type: 'string',
-                          format: 'json',
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-            { body: { a: '{ "z": "valid json" }', b: undefined } }
-          ).log.entries[0].request.postData.text
-        ).toBe(JSON.stringify({ a: { z: 'valid json' }, b: undefined }));
-      });
-
-      it('should leave user specified empty object JSON alone', () => {
-        expect(
-          oasToHar(
-            oas,
-            {
-              path: '/body',
-              method: 'post',
-              requestBody: {
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      required: ['a'],
-                      properties: {
-                        a: {
-                          type: 'string',
-                          format: 'json',
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-            { body: { a: '{}' } }
-          ).log.entries[0].request.postData.text
-        ).toBe(JSON.stringify({ a: {} }));
-      });
-    });
-  });
-
-  describe('formData values', () => {
-    it('should not add on empty unrequired values', () => {
-      expect(
-        oasToHar(oas, {
-          path: '/body',
-          method: 'get',
-          requestBody: {
-            content: {
-              'application/x-www-form-urlencoded': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    a: {
-                      type: 'string',
-                    },
-                  },
-                },
-              },
-            },
-          },
-        }).log.entries[0].request.postData
-      ).toBeUndefined();
-    });
-
-    // TODO extensions[SEND_DEFAULTS]
-    it.skip('should set defaults if no value provided but is required', () => {
-      expect(
-        oasToHar(oas, {
-          path: '/body',
-          method: 'get',
-          requestBody: {
-            content: {
-              'application/x-www-form-urlencoded': {
-                schema: {
-                  type: 'object',
-                  required: ['a'],
-                  properties: {
-                    a: {
-                      type: 'string',
-                    },
-                  },
-                },
-                example: { a: 'value' },
-              },
-            },
-          },
-        }).log.entries[0].request.postData.text
-      ).toBe(JSON.stringify({ a: 'value' }));
-    });
-
-    it('should not add undefined formData into postData', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/x-www-form-urlencoded': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      foo: {
-                        type: 'string',
-                      },
-                      bar: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { formData: { foo: undefined, bar: undefined } }
-        ).log.entries[0].request.postData
-      ).toBeUndefined();
-    });
-
-    it('should pass in value if one is set and prioritise provided values', () => {
-      expect(
-        oasToHar(
-          oas,
-          {
-            path: '/body',
-            method: 'get',
-            requestBody: {
-              content: {
-                'application/x-www-form-urlencoded': {
-                  schema: {
-                    type: 'object',
-                    required: ['a'],
-                    properties: {
-                      a: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-          { formData: { a: 'test', b: [1, 2, 3] } }
-        ).log.entries[0].request.postData.params
-      ).toStrictEqual([
-        { name: 'a', value: 'test' },
-        { name: 'b', value: '1,2,3' },
-      ]);
-    });
-
-    it('should support nested objects', () => {
-      // eslint-disable-next-line global-require
-      const spec = new Oas(require('./__fixtures__/formData-nested-object.json'));
-      const operation = spec.operation('/anything', 'post');
-      const formData = {
-        id: 12345,
-        Request: {
-          MerchantId: 'buster',
-        },
-      };
-
-      const har = oasToHar(spec, operation, { formData });
-      expect(har.log.entries[0].request.postData.params).toStrictEqual([
-        { name: 'id', value: 12345 },
-        { name: 'Request', value: '{"MerchantId":"buster"}' },
-      ]);
-    });
-  });
-});
-
-describe('common parameters', () => {
-  const operation = {
-    ...commonParameters.paths['/anything/{id}'].post,
-    path: '/anything/{id}',
-    method: 'post',
-  };
-
-  it('should work for common parameters', async () => {
-    const har = oasToHar(new Oas(commonParameters), operation, {
-      path: { id: 1234 },
-      header: { 'x-extra-id': 'abcd' },
-      query: { limit: 10 },
-      cookie: { authtoken: 'password' },
-    });
-
-    await expect(har).toBeAValidHAR();
-    expect(har.log.entries[0].request).toStrictEqual({
-      bodySize: 0,
-      cookies: [{ name: 'authtoken', value: 'password' }],
-      headers: [{ name: 'x-extra-id', value: 'abcd' }],
-      headersSize: 0,
-      httpVersion: 'HTTP/1.1',
-      queryString: [{ name: 'limit', value: '10' }],
-      method: 'POST',
-      url: 'https://httpbin.org/anything/1234',
-    });
-  });
-
-  it('should not mutate the original pathOperation that was passed in', () => {
-    const existingCount = operation.parameters.length;
-
-    oasToHar(new Oas(commonParameters), operation, {
-      path: { id: 1234 },
-      header: { 'x-extra-id': 'abcd' },
-      query: { limit: 10 },
-      cookie: { authtoken: 'password' },
-    });
-
-    expect(operation.parameters).toHaveLength(existingCount);
   });
 });
 
 describe('auth', () => {
   it('should work for header', () => {
-    expect(
-      oasToHar(
-        new Oas({
-          components: {
-            securitySchemes: {
-              'auth-header': {
-                type: 'apiKey',
-                name: 'x-auth-header',
-                in: 'header',
-              },
-            },
+    const spec = new Oas({
+      paths: {
+        '/security': {
+          get: {
+            security: [{ 'auth-header': [] }],
           },
-        }),
-        {
-          path: '/security',
-          method: 'get',
-          security: [{ 'auth-header': [] }],
         },
-        {},
-        {
-          'auth-header': 'value',
-        }
-      ).log.entries[0].request.headers
+      },
+      components: {
+        securitySchemes: {
+          'auth-header': {
+            type: 'apiKey',
+            name: 'x-auth-header',
+            in: 'header',
+          },
+        },
+      },
+    });
+
+    expect(
+      oasToHar(spec, spec.operation('/security', 'get'), {}, { 'auth-header': 'value' }).log.entries[0].request.headers
     ).toStrictEqual([
       {
         name: 'x-auth-header',
@@ -1668,24 +192,29 @@ describe('auth', () => {
   });
 
   it('should work for query', () => {
+    const spec = new Oas({
+      paths: {
+        '/security': {
+          get: {
+            security: [{ 'auth-query': [] }],
+          },
+        },
+      },
+      components: {
+        securitySchemes: {
+          'auth-query': {
+            type: 'apiKey',
+            name: 'authQuery',
+            in: 'query',
+          },
+        },
+      },
+    });
+
     expect(
       oasToHar(
-        new Oas({
-          components: {
-            securitySchemes: {
-              'auth-query': {
-                type: 'apiKey',
-                name: 'authQuery',
-                in: 'query',
-              },
-            },
-          },
-        }),
-        {
-          path: '/security',
-          method: 'get',
-          security: [{ 'auth-query': [] }],
-        },
+        spec,
+        spec.operation('/security', 'get'),
         {},
         {
           'auth-query': 'value',
@@ -1700,24 +229,29 @@ describe('auth', () => {
   });
 
   it('should work for cookie', () => {
+    const spec = new Oas({
+      paths: {
+        '/security': {
+          get: {
+            security: [{ 'auth-cookie': [] }],
+          },
+        },
+      },
+      components: {
+        securitySchemes: {
+          'auth-cookie': {
+            type: 'apiKey',
+            name: 'authCookie',
+            in: 'cookie',
+          },
+        },
+      },
+    });
+
     expect(
       oasToHar(
-        new Oas({
-          components: {
-            securitySchemes: {
-              'auth-cookie': {
-                type: 'apiKey',
-                name: 'authCookie',
-                in: 'cookie',
-              },
-            },
-          },
-        }),
-        {
-          path: '/security',
-          method: 'get',
-          security: [{ 'auth-cookie': [] }],
-        },
+        spec,
+        spec.operation('/security', 'get'),
         {},
         {
           'auth-cookie': 'value',
@@ -1732,29 +266,34 @@ describe('auth', () => {
   });
 
   it('should work for multiple (||)', () => {
+    const spec = new Oas({
+      paths: {
+        '/security': {
+          get: {
+            security: [{ 'auth-header': [] }, { 'auth-header2': [] }],
+          },
+        },
+      },
+      components: {
+        securitySchemes: {
+          'auth-header': {
+            type: 'apiKey',
+            name: 'x-auth-header',
+            in: 'header',
+          },
+          'auth-header2': {
+            type: 'apiKey',
+            name: 'x-auth-header2',
+            in: 'header',
+          },
+        },
+      },
+    });
+
     expect(
       oasToHar(
-        new Oas({
-          components: {
-            securitySchemes: {
-              'auth-header': {
-                type: 'apiKey',
-                name: 'x-auth-header',
-                in: 'header',
-              },
-              'auth-header2': {
-                type: 'apiKey',
-                name: 'x-auth-header2',
-                in: 'header',
-              },
-            },
-          },
-        }),
-        {
-          path: '/security',
-          method: 'get',
-          security: [{ 'auth-header': [] }, { 'auth-header2': [] }],
-        },
+        spec,
+        spec.operation('/security', 'get'),
         {},
         {
           'auth-header': 'value',
@@ -1774,29 +313,34 @@ describe('auth', () => {
   });
 
   it('should work for multiple (&&)', () => {
+    const spec = new Oas({
+      paths: {
+        '/security': {
+          get: {
+            security: [{ 'auth-header': [], 'auth-header2': [] }],
+          },
+        },
+      },
+      components: {
+        securitySchemes: {
+          'auth-header': {
+            type: 'apiKey',
+            name: 'x-auth-header',
+            in: 'header',
+          },
+          'auth-header2': {
+            type: 'apiKey',
+            name: 'x-auth-header2',
+            in: 'header',
+          },
+        },
+      },
+    });
+
     expect(
       oasToHar(
-        new Oas({
-          components: {
-            securitySchemes: {
-              'auth-header': {
-                type: 'apiKey',
-                name: 'x-auth-header',
-                in: 'header',
-              },
-              'auth-header2': {
-                type: 'apiKey',
-                name: 'x-auth-header2',
-                in: 'header',
-              },
-            },
-          },
-        }),
-        {
-          path: '/security',
-          method: 'get',
-          security: [{ 'auth-header': [], 'auth-header2': [] }],
-        },
+        spec,
+        spec.operation('/security', 'get'),
         {},
         {
           'auth-header': 'value',
@@ -1816,202 +360,48 @@ describe('auth', () => {
   });
 
   it('should not set non-existent values', () => {
-    expect(
-      oasToHar(
-        new Oas({
-          components: {
-            securitySchemes: {
-              'auth-header': {
-                type: 'apiKey',
-                name: 'x-auth-header',
-                in: 'header',
-              },
-            },
-          },
-        }),
-        {
-          path: '/security',
-          method: 'get',
-          security: [{ 'auth-header': [] }],
-        },
-        {},
-        {}
-      ).log.entries[0].request.headers
-    ).toStrictEqual([]);
-  });
-});
-
-describe('content-type & accept header', () => {
-  const operation = {
-    path: '/body',
-    method: 'get',
-    requestBody: {
-      content: {
-        'application/json': {
-          schema: {
-            type: 'object',
-            required: ['a'],
-            properties: {
-              a: {
-                type: 'string',
-              },
-            },
-          },
-          example: { a: 'value' },
-        },
-      },
-    },
-  };
-
-  it('should be sent through if there are no body values but there is a requestBody', async () => {
-    let har = oasToHar(oas, operation, {});
-
-    await expect(har).toBeAValidHAR();
-    expect(har.log.entries[0].request.headers).toStrictEqual([{ name: 'Content-Type', value: 'application/json' }]);
-
-    har = oasToHar(oas, operation, { query: { a: 1 } });
-    await expect(har).toBeAValidHAR();
-    expect(har.log.entries[0].request.headers).toStrictEqual([{ name: 'Content-Type', value: 'application/json' }]);
-  });
-
-  it('should be sent through if there are any body values', async () => {
-    const har = oasToHar(oas, operation, { body: { a: 'test' } });
-
-    await expect(har).toBeAValidHAR();
-    expect(har.log.entries[0].request.headers).toStrictEqual([{ name: 'Content-Type', value: 'application/json' }]);
-  });
-
-  it('should be sent through if there are any formData values', async () => {
-    const har = oasToHar(oas, operation, { formData: { a: 'test' } });
-    await expect(har).toBeAValidHAR();
-    expect(har.log.entries[0].request.headers).toStrictEqual([{ name: 'Content-Type', value: 'application/json' }]);
-  });
-
-  it('should fetch the type from the first `requestBody.content` and first `responseBody.content` object', async () => {
-    const har = oasToHar(
-      oas,
-      {
-        path: '/body',
-        method: 'get',
-        requestBody: {
-          content: {
-            'text/xml': {
-              schema: {
-                type: 'object',
-                required: ['a'],
-                properties: {
-                  a: {
-                    type: 'string',
-                  },
-                },
-              },
-              example: { a: 'value' },
-            },
+    const spec = new Oas({
+      paths: {
+        '/security': {
+          get: {
+            security: [{ 'auth-header': [] }],
           },
         },
       },
-      { body: { a: 'test' } }
-    );
-
-    await expect(har).toBeAValidHAR();
-    expect(har.log.entries[0].request.headers).toStrictEqual([{ name: 'Content-Type', value: 'text/xml' }]);
-    expect(har.log.entries[0].request.postData.mimeType).toBe('text/xml');
-  });
-
-  // Whether this is right or wrong, i'm not sure but this is what readme currently does
-  it('should prioritise json if it exists', async () => {
-    const har = oasToHar(
-      oas,
-      {
-        path: '/body',
-        method: 'get',
-        requestBody: {
-          content: {
-            'text/xml': {
-              schema: {
-                type: 'string',
-                required: ['a'],
-                properties: {
-                  a: {
-                    type: 'string',
-                  },
-                },
-              },
-              example: { a: 'value' },
-            },
-            'application/json': {
-              schema: {
-                type: 'object',
-                required: ['a'],
-                properties: {
-                  a: {
-                    type: 'string',
-                  },
-                },
-              },
-              example: { a: 'value' },
-            },
+      components: {
+        securitySchemes: {
+          'auth-header': {
+            type: 'apiKey',
+            name: 'x-auth-header',
+            in: 'header',
           },
         },
       },
-      { body: { a: 'test' } }
-    );
+    });
 
-    await expect(har).toBeAValidHAR();
-    expect(har.log.entries[0].request.headers).toStrictEqual([{ name: 'Content-Type', value: 'application/json' }]);
-  });
-
-  it("should only add a content-type if one isn't already present", async () => {
-    const har = oasToHar(
-      new Oas({
-        'x-headers': [{ key: 'Content-Type', value: 'multipart/form-data' }],
-      }),
-      {
-        path: '/',
-        method: 'put',
-        requestBody: {
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                required: ['a'],
-                properties: {
-                  a: {
-                    type: 'string',
-                  },
-                },
-              },
-              example: { a: 'value' },
-            },
-          },
-        },
-      },
-      { body: { a: 'test' } }
-    );
-
-    await expect(har).toBeAValidHAR();
-
-    // `Content-Type: application/json` would normally appear here if there were no `x-headers`, but since there is
-    // we should default to that so as to we don't double up on Content-Type headers.
-    expect(har.log.entries[0].request.headers).toStrictEqual([{ name: 'Content-Type', value: 'multipart/form-data' }]);
-
-    expect(har.log.entries[0].request.postData.mimeType).toBe('multipart/form-data');
+    const har = oasToHar(spec, spec.operation('/security', 'get'), {}, {});
+    expect(har.log.entries[0].request.headers).toStrictEqual([]);
   });
 });
 
 describe('x-headers', () => {
   it('should append any static headers to the request', () => {
-    expect(
-      oasToHar(
-        new Oas({
-          'x-headers': [
-            {
-              key: 'x-api-key',
-              value: '123456',
-            },
-          ],
-        })
-      ).log.entries[0].request.headers
-    ).toStrictEqual([{ name: 'x-api-key', value: '123456' }]);
+    const spec = new Oas({
+      paths: {
+        '/': {
+          post: {},
+        },
+      },
+      'x-headers': [
+        {
+          key: 'x-api-key',
+          value: '123456',
+        },
+      ],
+    });
+
+    expect(oasToHar(spec, spec.operation('/', 'post')).log.entries[0].request.headers).toStrictEqual([
+      { name: 'x-api-key', value: '123456' },
+    ]);
   });
 });
