+++ package.json
@@ -23,7 +23,7 @@
   },
   "dependencies": {
     "@readme/oas-extensions": "^14.0.0",
-    "oas": "^17.1.0",
+    "oas": "^17.3.2",
     "parse-data-url": "^4.0.1"
   },
   "devDependencies": {
